Merge from mainline

+++ src/main/org/apache/tools/ant/taskdefs/WaitFor.java
@@ -244,21 +244,21 @@ public class WaitFor extends ConditionBase {
 
         /** Constructor the Unit enumerated type. */
         public Unit() {
-            timeTable.put(MILLISECOND, Long.valueOf(1L));
-            timeTable.put(SECOND,      Long.valueOf(ONE_SECOND));
-            timeTable.put(MINUTE,      Long.valueOf(ONE_MINUTE));
-            timeTable.put(HOUR,        Long.valueOf(ONE_HOUR));
-            timeTable.put(DAY,         Long.valueOf(ONE_DAY));
-            timeTable.put(WEEK,        Long.valueOf(ONE_WEEK));
+            timeTable.put(MILLISECOND, 1L);
+            timeTable.put(SECOND, ONE_SECOND);
+            timeTable.put(MINUTE, ONE_MINUTE);
+            timeTable.put(HOUR, ONE_HOUR);
+            timeTable.put(DAY, ONE_DAY);
+            timeTable.put(WEEK, ONE_WEEK);
         }
 
         /**
-         * Convert the value to a multipler (millisecond to unit).
-         * @return a multipler (a long value)
+         * Convert the value to a multiplier (millisecond to unit).
+         * @return a multiplier (a long value)
          */
         public long getMultiplier() {
             String key = getValue().toLowerCase(Locale.ENGLISH);
-            return timeTable.get(key).longValue();
+            return timeTable.get(key);
         }
 
         /**
